Clean up angle definitions
From now on, every angular value must be an implemention of AngleType
interface. It includes the averages of angles or the pseudophase pucker angle
which are not based on specific atoms.

This change removes references to angles by String (by their name) and also
removes all the hardcoded names that were spread across the source code base